Move dialog details of implementation to the internal folder
