+++ src/Avalonia.Dialogs/Internal/FileSizeStringConverter.cs
@@ -1,10 +1,8 @@
-using Avalonia.Data.Converters;
 using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Text;
+using Avalonia.Data.Converters;
 
-namespace Avalonia.Dialogs
+namespace Avalonia.Dialogs.Internal
 {
     public class FileSizeStringConverter : IValueConverter
     {
