+++ src/Avalonia.Dialogs/Internal/InternalViewModelBase.cs
