+++ src/Avalonia.Dialogs/Internal/ManagedFileChooserItemType.cs
@@ -1,4 +1,4 @@
-namespace Avalonia.Dialogs
+namespace Avalonia.Dialogs.Internal
 {
     public enum ManagedFileChooserItemType
     {
