+++ src/Avalonia.Dialogs/Internal/ManagedFileChooserSources.cs
@@ -2,13 +2,10 @@ using System;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
-using System.Reactive.Linq;
-using System.Runtime.InteropServices;
 using Avalonia.Controls.Platform;
-using Avalonia.Threading;
 using Avalonia.Utilities;
 
-namespace Avalonia.Dialogs
+namespace Avalonia.Dialogs.Internal
 {
     internal class ManagedFileChooserSources
     {
